feat: HistoryDetailScreen 화면 구현

+++ domain/src/main/java/com/analysis/domain/model/VerificationResult.kt
@@ -5,9 +5,9 @@ import java.time.LocalDateTime
 data class VerificationResult(
     val id:String,
     val title:String,
-    val similarity:Double,
-    val pressure:Double,
-    val inclination:Double,
+    val similarity:Float,
+    val pressure:Float,
+    val inclination:Float,
     val verificationImgUrl: String,
     val createdAt: LocalDateTime,
 )
